// src/api/back/models/MatchGameDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    MatchStatus,
    MatchStatusFromJSON,
    MatchStatusFromJSONTyped,
    MatchStatusToJSON,
    ParticipantResultDto,
    ParticipantResultDtoFromJSON,
    ParticipantResultDtoFromJSONTyped,
    ParticipantResultDtoToJSON,
} from './';

/**
 * 
 * @export
 * @interface MatchGameDto
 */
export interface MatchGameDto {
    /**
     * 
     * @type {MatchStatus}
     * @memberof MatchGameDto
     */
    status: MatchStatus;
    /**
     * 
     * @type {string}
     * @memberof MatchGameDto
     */
    id: string;
    /**
     * 
     * @type {number}
     * @memberof MatchGameDto
     */
    bracketMatchId: number;
    /**
     * 
     * @type {number}
     * @memberof MatchGameDto
     */
    externalMatchId?: number;
    /**
     * 
     * @type {string}
     * @memberof MatchGameDto
     */
    scheduledDate?: string;
    /**
     * 
     * @type {number}
     * @memberof MatchGameDto
     */
    teamOffset: number;
    /**
     * 
     * @type {number}
     * @memberof MatchGameDto
     */
    number: number;
    /**
     * 
     * @type {boolean}
     * @memberof MatchGameDto
     */
    finished: boolean;
    /**
     * 
     * @type {ParticipantResultDto}
     * @memberof MatchGameDto
     */
    opponent1?: ParticipantResultDto;
    /**
     * 
     * @type {ParticipantResultDto}
     * @memberof MatchGameDto
     */
    opponent2?: ParticipantResultDto;
}

export function MatchGameDtoFromJSON(json: any): MatchGameDto {
    return MatchGameDtoFromJSONTyped(json, false);
}

export function MatchGameDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): MatchGameDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'status': MatchStatusFromJSON(json['status']),
        'id': json['id'],
        'bracketMatchId': json['bracketMatchId'],
        'externalMatchId': !exists(json, 'externalMatchId') ? undefined : json['externalMatchId'],
        'scheduledDate': !exists(json, 'scheduledDate') ? undefined : json['scheduledDate'],
        'teamOffset': json['teamOffset'],
        'number': json['number'],
        'finished': json['finished'],
        'opponent1': !exists(json, 'opponent1') ? undefined : ParticipantResultDtoFromJSON(json['opponent1']),
        'opponent2': !exists(json, 'opponent2') ? undefined : ParticipantResultDtoFromJSON(json['opponent2']),
    };
}

export function MatchGameDtoToJSON(value?: MatchGameDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'status': MatchStatusToJSON(value.status),
        'id': value.id,
        'bracketMatchId': value.bracketMatchId,
        'externalMatchId': value.externalMatchId,
        'scheduledDate': value.scheduledDate,
        'teamOffset': value.teamOffset,
        'number': value.number,
        'finished': value.finished,
        'opponent1': ParticipantResultDtoToJSON(value.opponent1),
        'opponent2': ParticipantResultDtoToJSON(value.opponent2),
    };
}